[front] chore(mcp): move MCP types to a specific files
