+++ front/components/spaces/SpaceSideBarMenu.tsx
@@ -15,7 +15,7 @@ import { Fragment, useCallback, useEffect, useMemo, useState } from "react";
 import { useTheme } from "@app/components/sparkle/ThemeContext";
 import { usePersistedNavigationSelection } from "@app/hooks/usePersistedNavigationSelection";
 import { MCP_SERVER_ICONS } from "@app/lib/actions/mcp_icons";
-import type { MCPServerViewType } from "@app/lib/actions/mcp_metadata";
+import type { MCPServerViewType } from "@app/lib/api/mcp";
 import { getConnectorProviderLogoWithFallback } from "@app/lib/connector_providers";
 import { getVisualForDataSourceViewContentNode } from "@app/lib/content_nodes";
 import { getDataSourceNameFromView } from "@app/lib/data_sources";
