+++ front/lib/actions/mcp_internal_actions/servers/data_source_utils.ts
@@ -6,7 +6,7 @@ import {
   ConfigurableToolInputSchemas,
   DATA_SOURCE_CONFIGURATION_URI_PATTERN,
 } from "@app/lib/actions/mcp_internal_actions/input_schemas";
-import type { InternalMCPServerDefinitionType } from "@app/lib/actions/mcp_metadata";
+import type { InternalMCPServerDefinitionType } from "@app/lib/api/mcp";
 import { AgentDataSourceConfiguration } from "@app/lib/models/assistant/actions/data_sources";
 import { DataSourceModel } from "@app/lib/resources/storage/models/data_source";
 import { getResourceNameAndIdFromSId } from "@app/lib/resources/string_ids";
