+++ front/lib/actions/mcp_internal_actions/servers/github.ts
@@ -3,7 +3,7 @@ import { Octokit } from "@octokit/core";
 import { z } from "zod";
 
 import { getAccessTokenForInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/authentication";
-import type { InternalMCPServerDefinitionType } from "@app/lib/actions/mcp_metadata";
+import type { InternalMCPServerDefinitionType } from "@app/lib/api/mcp";
 import type { Authenticator } from "@app/lib/auth";
 import { normalizeError } from "@app/types";
