+++ front/lib/swr/mcp_server_views.ts
@@ -2,10 +2,7 @@ import { useSendNotification } from "@dust-tt/sparkle";
 import { useCallback, useMemo } from "react";
 import type { Fetcher } from "swr";
 
-import type {
-  MCPServerType,
-  MCPServerViewType,
-} from "@app/lib/actions/mcp_metadata";
+import type { MCPServerType, MCPServerViewType } from "@app/lib/api/mcp";
 import { useMCPServers } from "@app/lib/swr/mcp_servers";
 import { fetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
 import type { DeleteMCPServerResponseBody } from "@app/pages/api/w/[wId]/mcp/[serverId]";
