+++ front/pages/api/w/[wId]/mcp/index.ts
@@ -4,12 +4,9 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { internalMCPServerNameToSId } from "@app/lib/actions/mcp_helper";
 import { isInternalMCPServerName } from "@app/lib/actions/mcp_internal_actions/constants";
-import type {
-  MCPServerType,
-  MCPServerTypeWithViews,
-} from "@app/lib/actions/mcp_metadata";
 import { fetchRemoteServerMetaDataByURL } from "@app/lib/actions/mcp_metadata";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
+import type { MCPServerType, MCPServerTypeWithViews } from "@app/lib/api/mcp";
 import type { Authenticator } from "@app/lib/auth";
 import { InternalMCPServerInMemoryResource } from "@app/lib/resources/internal_mcp_server_in_memory_resource";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
