+++ front/pages/w/[wId]/actions.tsx
@@ -6,8 +6,8 @@ import { AdminActionsList } from "@app/components/actions/mcp/ActionsList";
 import { MCPServerDetails } from "@app/components/actions/mcp/MCPServerDetails";
 import { subNavigationAdmin } from "@app/components/navigation/config";
 import AppLayout from "@app/components/sparkle/AppLayout";
-import type { MCPServerType } from "@app/lib/actions/mcp_metadata";
 import { ACTION_SPECIFICATIONS } from "@app/lib/actions/utils";
+import type { MCPServerType } from "@app/lib/api/mcp";
 import { getFeatureFlags } from "@app/lib/auth";
 import { withDefaultUserAuthPaywallWhitelisted } from "@app/lib/iam/session";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
