+++ front/tests/utils/RemoteMCPServerFactory.ts
@@ -1,6 +1,6 @@
 import { faker } from "@faker-js/faker";
 
-import type { MCPToolType } from "@app/lib/actions/mcp_metadata";
+import type { MCPToolType } from "@app/lib/api/mcp";
 import { Authenticator } from "@app/lib/auth";
 import { RemoteMCPServerResource } from "@app/lib/resources/remote_mcp_servers_resource";
 import type { WorkspaceType } from "@app/types";
